fix(lib): inline zip creation
We previously used an execSync to use a promise based API in a synchronous
environment and due to a misunderstanding in how bundledDependencies can be
used.

Closes #3176
