+++ packages/@cdktf/commons/ambient.d.ts
@@ -10,3 +10,5 @@ declare module 'child_process' {
 }
 
 declare module 'tunnel-agent';
+
+declare module "archiver";
