+++ packages/@cdktf/commons/tsconfig.json
@@ -38,6 +38,7 @@
   "exclude": [
     "node_modules",
     "../../node_modules/**",
+    "../../cdktf/**",
     "**/*.d.ts",
     "**/*.test.tsx"
   ],
