+++ packages/@cdktf/provider-generator/ambient.d.ts
@@ -10,3 +10,4 @@ declare module 'child_process' {
 }
 
 declare module 'tunnel-agent';
+declare module "archiver";
