+++ packages/@cdktf/provider-schema/ambient.d.ts
@@ -6,3 +6,5 @@ declare module 'child_process' {
         signal?: AbortSignal;
     }
 }
+
+declare module "archiver";
