+++ packages/cdktf/lib/ambient.d.ts
@@ -1,3 +1,3 @@
 // Copyright (c) HashiCorp, Inc
 // Licensed under the Mozilla Public License 2.0
-declare module "archiver";
+// declare module "archiver";
