+++ packages/cdktf/package.json
@@ -104,6 +104,7 @@
   },
   "license": "MPL-2.0",
   "devDependencies": {
+    "@types/archiver": "6.0.2",
     "@types/jest": "29.5.11",
     "@types/json-stable-stringify": "1.0.36",
     "@types/minimatch": "5.1.2",
@@ -123,7 +124,7 @@
     "typescript": "5.2.2"
   },
   "dependencies": {
-    "archiver": "5.3.2",
+    "archiver": "6.0.1",
     "json-stable-stringify": "1.1.0",
     "semver": "7.5.4"
   },
